--- CESA-2020-Prototype/Assets/DebugCamera.cs
@@ -1,22 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class DebugCamera : MonoBehaviour
-{
-    [SerializeField]
-    CameraShake cameraShake;
-    // Start is called before the first frame update
-    void Start()
-    {
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        if(Input.GetKeyDown(KeyCode.A))
-        {
-            cameraShake.Shake(0.3f, 0.2f);
-        }
-    }
-}
