+++ CESA-2020-Prototype/Assets/Scripts/CameraController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c28accce1bc44e34cb02307b1dd652ce
+guid: 54437d1cea1698a43aaa6224cccd16d7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
